+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed `Clone` impl for `Node`.
+- Removed `Nodes::iter_mut` method.
 
 [0.4.1] - 2023-05-25
 --------------------